+++ src/reducer/game/game.test.js
@@ -3,11 +3,8 @@ import {
   ActionType,
   isArtistAnswerCorrect,
   isGenreAnswerCorrect,
-  Operation,
   reducer,
-} from './reducer';
-import {createAPI} from "./api";
-import MockAdapter from "axios-mock-adapter";
+} from './game';
 
 
 describe(`Business logic is correct`, () => {
@@ -243,7 +240,6 @@ describe(`Reducer works correctly`, () => {
     expect(reducer(undefined, {})).toEqual({
       step: -1,
       mistakes: 0,
-      questions: [],
     });
   });
 
@@ -307,25 +303,5 @@ describe(`Reducer works correctly`, () => {
       mistakes: 0,
       questions: [`asdf`],
     });
-  });
-
-  it(`Should make a correct API call to /questions`, function () {
-    const dispatch = jest.fn();
-    const api = createAPI(dispatch);
-    const apiMock = new MockAdapter(api);
-    const questionLoader = Operation.loadQuestions();
-
-    apiMock
-      .onGet(`/questions`)
-      .reply(200, [{fake: true}]);
-
-    return questionLoader(dispatch, jest.fn(), api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionType.LOAD_QUESTIONS,
-          payload: [{fake: true}],
-        });
-      });
   });
 });
